first attempt at putting all the pieces together

+++ 5_Project3B/draw_bounding_box.py
@@ -6,6 +6,6 @@ import cv2
 
 def draw_bounding_box(bb_corners, img):
     bb_img = img
-    x, y, w, h = cv2.boundingRect(bb_corners)
+    x, y, w, h = cv2.boundingRect(bb_corners.astype(int))
     cv2.rectangle(bb_img, (x, y), (x+w, y+h), (0, 255, 0), 2)
     return bb_img
